+++ app/src/main/kotlin/ac/mdiq/podcini/receiver/ConnectivityActionReceiver.kt
@@ -1,9 +1,8 @@
 package ac.mdiq.podcini.receiver
 
+import ac.mdiq.podcini.PodciniApp.Companion.getApp
 import ac.mdiq.podcini.automation.autodownload
 import ac.mdiq.podcini.net.download.service.DownloadServiceInterface
-import ac.mdiq.podcini.net.utils.NetworkUtils.networkAllowAutoDownload
-import ac.mdiq.podcini.net.utils.NetworkUtils.isNetworkRestricted
 import ac.mdiq.podcini.utils.Logd
 import ac.mdiq.podcini.utils.Logt
 import ac.mdiq.podcini.config.ClientConfigurator
@@ -24,12 +23,12 @@ class ConnectivityActionReceiver : BroadcastReceiver() {
     }
 
     private fun networkChangedDetected(context: Context) {
-        if (networkAllowAutoDownload) {
+        if (getApp().networkMonitor.networkAllowAutoDownload) {
             Logd(TAG, "auto-dl network available, starting auto-download")
             autodownload(context)
         } else { // if new network is Wi-Fi, finish ongoing downloads,
             // otherwise cancel all downloads
-            if (isNetworkRestricted) {
+            if (getApp().networkMonitor.isNetworkRestricted) {
                 Logt(TAG, "Device is no longer connected to Wi-Fi. Cancelling ongoing downloads")
                 DownloadServiceInterface.impl?.cancelAll(context)
             }