+++ app/src/main/kotlin/ac/mdiq/podcini/storage/model/Volume.kt
@@ -1,5 +1,6 @@
 package ac.mdiq.podcini.storage.model
 
+import ac.mdiq.podcini.storage.database.realm
 import io.github.xilinjia.krdb.types.RealmObject
 import io.github.xilinjia.krdb.types.annotations.PrimaryKey
 
@@ -12,4 +13,14 @@ class Volume : RealmObject {
     var uriString: String = ""
 
     var parentId: Long = -1L
+}
+
+fun Volume.getAllChildren(): List<Volume> {
+    val result = mutableListOf<Volume>()
+    val subVolumes = realm.query(Volume::class).query("parentId == $id").find()
+    for (child in subVolumes) {
+        result.add(child)
+        result.addAll(child.getAllChildren())
+    }
+    return result
 }